Move MarkdownPost class into Models\Pages namespace
