+++ packages/framework/resources/views/components/blog-post-feed.blade.php
@@ -1,3 +1,3 @@
-@foreach(\Hyde\Framework\Models\MarkdownPost::getLatestPosts() as $post)
-	@include('hyde::components.article-excerpt')
+@foreach(\Hyde\Framework\Models\Pages\MarkdownPost::getLatestPosts() as $post)
+    @include('hyde::components.article-excerpt')
 @endforeach
