+++ packages/framework/src/Services/CollectionService.php
@@ -7,7 +7,7 @@ use Hyde\Framework\Hyde;
 use Hyde\Framework\Models\BladePage;
 use Hyde\Framework\Models\DocumentationPage;
 use Hyde\Framework\Models\Pages\MarkdownPage;
-use Hyde\Framework\Models\MarkdownPost;
+use Hyde\Framework\Models\Pages\MarkdownPost;
 
 /**
  * Contains service methods to return helpful collections of arrays and lists.
