+++ packages/framework/src/Services/RssFeedService.php
@@ -4,7 +4,7 @@ namespace Hyde\Framework\Services;
 
 use Hyde\Framework\Helpers\Features;
 use Hyde\Framework\Hyde;
-use Hyde\Framework\Models\MarkdownPost;
+use Hyde\Framework\Models\Pages\MarkdownPost;
 use SimpleXMLElement;
 
 /**
