+++ packages/framework/tests/Feature/Commands/StaticSiteBuilderPostModuleTest.php
@@ -3,7 +3,7 @@
 namespace Hyde\Framework\Testing\Feature\Commands;
 
 use Hyde\Framework\Hyde;
-use Hyde\Framework\Models\MarkdownPost;
+use Hyde\Framework\Models\Pages\MarkdownPost;
 use Hyde\Framework\StaticPageBuilder;
 use Hyde\Testing\TestCase;
