+++ packages/framework/tests/Feature/Services/CollectionServiceTest.php
@@ -5,7 +5,7 @@ namespace Hyde\Framework\Testing\Feature\Services;
 use Hyde\Framework\Hyde;
 use Hyde\Framework\Models\DocumentationPage;
 use Hyde\Framework\Models\Pages\MarkdownPage;
-use Hyde\Framework\Models\MarkdownPost;
+use Hyde\Framework\Models\Pages\MarkdownPost;
 use Hyde\Framework\Services\CollectionService;
 use Hyde\Testing\TestCase;
 use Illuminate\Support\Facades\File;
