+++ packages/framework/tests/Feature/StaticPageBuilderTest.php
@@ -7,7 +7,7 @@ use Hyde\Framework\HydeServiceProvider;
 use Hyde\Framework\Models\BladePage;
 use Hyde\Framework\Models\DocumentationPage;
 use Hyde\Framework\Models\Pages\MarkdownPage;
-use Hyde\Framework\Models\MarkdownPost;
+use Hyde\Framework\Models\Pages\MarkdownPost;
 use Hyde\Framework\StaticPageBuilder;
 use Hyde\Testing\ResetsApplication;
 use Hyde\Testing\TestCase;
