+++ packages/framework/tests/Unit/MarkdownPageModelConstructorArgumentsAreOptionalTest.php
@@ -4,7 +4,7 @@ namespace Hyde\Framework\Testing\Unit;
 
 use Hyde\Framework\Models\DocumentationPage;
 use Hyde\Framework\Models\Pages\MarkdownPage;
-use Hyde\Framework\Models\MarkdownPost;
+use Hyde\Framework\Models\Pages\MarkdownPost;
 use Hyde\Testing\TestCase;
 
 class MarkdownPageModelConstructorArgumentsAreOptionalTest extends TestCase
