+++ packages/framework/tests/Unit/SourceFilesInCustomDirectoriesCanBeCompiledTest.php
@@ -6,7 +6,7 @@ use Hyde\Framework\Hyde;
 use Hyde\Framework\Models\BladePage;
 use Hyde\Framework\Models\DocumentationPage;
 use Hyde\Framework\Models\Pages\MarkdownPage;
-use Hyde\Framework\Models\MarkdownPost;
+use Hyde\Framework\Models\Pages\MarkdownPost;
 use Hyde\Framework\Services\DiscoveryService;
 use Hyde\Framework\StaticPageBuilder;
 use Hyde\Testing\TestCase;
